have a shared/context helpers
