+++ src/page/bell/AnimatedElement.ts
@@ -1,7 +1,8 @@
+import { containsMatch } from 'src/shared/context';
 import { addCssClass, removeCssClass } from 'src/shared/helpers/dom';
 import Log from '../../shared/libraries/Log';
 import OneSignalEvent from '../../shared/services/OneSignalEvent';
-import { contains, once } from '../../shared/utils/utils';
+import { once } from '../../shared/utils/utils';
 
 export default class AnimatedElement {
   public selector: string;
@@ -74,7 +75,7 @@ export default class AnimatedElement {
             (event: Event, destroyListenerFn: () => void) => {
               if (
                 event.target === this.element &&
-                contains(
+                containsMatch(
                   this.targetTransitionEvents,
                   (event as any).propertyName,
                 )
@@ -128,7 +129,7 @@ export default class AnimatedElement {
               }, this.transitionCheckTimeout);
               if (
                 event.target === this.element &&
-                contains(
+                containsMatch(
                   this.targetTransitionEvents,
                   (event as any).propertyName,
                 )
