+++ src/shared/utils/OneSignalUtils.ts
@@ -1,4 +1,4 @@
-import { Utils } from '../context/Utils';
+import { stringify } from '../context';
 import { isBrowser } from '../environment/environment';
 import Log from '../libraries/Log';
 
@@ -19,9 +19,7 @@ export class OneSignalUtils {
   }
 
   public static logMethodCall(methodName: string, ...args: any[]) {
-    return Log.debug(
-      `Called ${methodName}(${args.map(Utils.stringify).join(', ')})`,
-    );
+    return Log.debug(`Called ${methodName}(${args.map(stringify).join(', ')})`);
   }
 
   static isSafari(): boolean {
